Fix uint8 overflow in sparse one-hot dot product for MSAs > 255 positions
get_onehotmsa_sparse used np.uint8 for sparse matrix data, causing the
dot product in weight methods v4/v5/v6 to silently wrap around when
position counts exceed 255. Changed to int16 which supports up to 32767.

+++ src/mysca/preprocess.py
@@ -561,7 +561,7 @@ def get_onehotmsa_sparse(msa, num_aa, gap):
     pos = np.tile(np.arange(npos, dtype=np.uint16), nseqs)
     mask = a < num_aa
     cols = pos[mask] * num_aa + a[mask]
-    data = np.ones(cols.shape[0], dtype=np.uint8)
+    data = np.ones(cols.shape[0], dtype=np.int16)
 
     onehotmsa = sp.csr_matrix(
         (data, (rows[mask], cols)),
